server/pyproject.toml
[tool.poetry]
name = "alix-webscraper"
version = "0.1.0"
description = ""
authors = ["daria425 <darianaumova5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.21.0"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.2"
fastapi = "^0.111.0"
gunicorn = "^22.0.0"
requests-html = "^0.10.0"
pymongo = {extras = ["srv"], version = "^4.7.3"}
webdriver-manager = "^4.0.1"
transformers = {extras = ["torch"], version = "^4.43.3"}
google-cloud-aiplatform = "^1.60.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
